==> what-is-my-quantum-computer-good-for/paper-simulations-fidelity/h-simulations/ring_processor/7-data-analysis.ipynb <==
import numpy as np
import pandas as pd

from tensorflow.keras.losses import BinaryCrossentropy

# a is the actual values while b is the predictions
def bce(a, b):
    return BinaryCrossentropy(from_logits = False)(a,b).numpy()

dtypes = ['train', 'validate', 'test']

experiments = [0, 1, 2, 3, 4]
exp_path = './experiment_{}/'
pred_path = exp_path + 'predictions/'
sps_path = exp_path + 'processed_inputs_and_outputs/'

metric_functions = {'r': lambda x, y: np.corrcoef(x, y)[0,1],
                    'kl': lambda x, y: bce(x, y) - bce(x, x),
                    'l1': lambda x, y: np.mean(np.abs(x - y))
                    }
model_predictions = {'cnn': 'processed_cnn_predictions.npz', 'custom-nn': 'processed_custom_neural_net_predictions.npz',
                    'mirrored-cnn': 'processed_cnn_mirrored_predictions.npz', 'mirrored-custom-nn': 'processed_custom_neural_net_mirrored_predictions.npz'}

# Computing MAE performance

def aggregate_data(metric_function, model_types, circuits):
    device_results = {}
    for exp in experiments:
        device_results[exp] = {}
        if circuits != 'mirrored':
            infidelities = np.load(sps_path.format(exp) + 'processed_infidelities.npz')
        else:
            infidelities = np.load(sps_path.format(exp) + 'processed_mirrored_infidelities.npz')
        for model_type in model_types:
            predictions = np.load(pred_path.format(exp)+model_predictions[model_type])
            if model_type[0] != 'm':
                predictions = {dt: predictions[dt].reshape((-1,)) for dt in dtypes}
            else:
                predictions = predictions['predictions'].reshape((-1,))
            if model_type in ['custom-nn', 'cnn']:
                device_results[exp][model_type] = metric_function(infidelities[circuits].reshape((-1,))/100, predictions[circuits].reshape((-1,))/100)
            elif model_type in ['mirrored-custom-nn', 'mirrored-cnn']:
                device_results[exp][model_type] = metric_function(infidelities['infidelities'].reshape((-1,))/100, predictions.reshape((-1,))/100)
    return device_results

def compute_diffs(results: dict, model1: str, model2: str) -> dict:
    '''
    Loop over the entries in results and compute the difference between model1's entries and model2's entries.
    '''
    diff_dict = {exp: result[model1] - result[model2] for exp, result in results.items()}
    return diff_dict

def compute_relative_diffs(results: dict, model1: str, model2: str) -> dict:
    '''
    Loop over the entries in results and compute the difference between model1's entries and model2's entries.
    '''
    diff_dict = {exp: 100*(result[model1] - result[model2]) / result[model2] for exp, result in results.items()}
    return diff_dict

def compute_statistic(results: dict, models: list, statistic: np.mean, s_kwargs = {}) -> dict: 
    '''
    Loop over the devices in results and compute the (statistic) value of a metric for a model type.
    '''
    r_dict = {model: [] for model in models}
    mean_dict = {model: None for model in models}
    for _, result in results.items():
        for model in models:
            r_dict[model].append(result[model])
    for model, item in r_dict.items():
        mean_dict[model] = statistic(item, **s_kwargs)
    return mean_dict

Comparing on random circuits

model_types = ['custom-nn', 'cnn']
circuits = 'test'

m1, m2 = model_types
metrics = ['l1']
for metric in metrics:
    results = aggregate_data(metric_functions[metric], model_types, circuits)

m_dict = compute_statistic(results, [m1, m2], np.mean)
std_dict = compute_statistic(results, [m1, m2], np.std, {'ddof': 1})
results, m_dict, std_dict

percent_diff_dict = compute_relative_diffs(results, m1, m2)
mean_percent_diff = np.mean([item for _, item in percent_diff_dict.items()])
std_percent_diff = np.std([item for _, item in percent_diff_dict.items()], ddof = 1)
mean_percent_diff, std_percent_diff

Comparing on mirror circuits

model_types = ['mirrored-custom-nn', 'mirrored-cnn']
circuits = 'mirrored'

metrics = ['l1']
for metric in metrics:
    results = aggregate_data(metric_functions[metric], model_types, circuits)

m1, m2 = model_types

m_dict = compute_statistic(results, [m1, m2], np.mean)
std_dict = compute_statistic(results, [m1, m2], np.std, {'ddof': 1})
results, m_dict, std_dict

percent_diff_dict = compute_relative_diffs(results, m1, m2)
mean_percent_diff = np.mean([item for _, item in percent_diff_dict.items()])
std_percent_diff = np.std([item for _, item in percent_diff_dict.items()], ddof = 1)
mean_percent_diff, std_percent_diff

# r value

Custom networks - random circuits

exps = [0, 1, 2, 3, 4]
ground = 'processed_infidelities.npz'
preds = 'processed_custom_neural_net_predictions.npz'
metric = 'r'

predictions = [np.load(pred_path.format(exp) + preds)['test'].reshape((-1,)) for exp in exps]
infidelities = [np.load(sps_path.format(exp) + ground)['test'].reshape((-1,)) for exp in exps]

metrics = [metric_functions[metric](p, i) for p,i in zip(predictions, infidelities)]
metrics, np.mean(metrics), np.std(metrics, ddof = 1)

CNNs - random circuits

exps = [0, 1, 2, 3, 4]
ground = 'processed_infidelities.npz'
preds = 'processed_cnn_predictions.npz'
metric = 'r'

predictions = [np.load(pred_path.format(exp) + preds)['test'].reshape((-1,)) for exp in exps]
infidelities = [np.load(sps_path.format(exp) + ground)['test'].reshape((-1,)) for exp in exps]

metrics = [metric_functions[metric](p, i) for p,i in zip(predictions, infidelities)]
metrics, np.mean(metrics), np.std(metrics)

Custom networks - mirrored circuits

exps = [0, 1, 2, 3, 4]
ground = 'processed_mirrored_infidelities.npz'
preds = 'processed_custom_neural_net_mirrored_predictions.npz'
metric = 'r'

predictions = [np.load(pred_path.format(exp) + preds)['predictions'] for exp in exps]
infidelities = [np.load(sps_path.format(exp) + ground)['infidelities'] for exp in exps]

metrics = [metric_functions[metric](p, i) for p,i in zip(predictions, infidelities)]
metrics, np.mean(metrics), np.std(metrics, ddof = 1)

CNNs - mirrored circuits

exps = [0, 1, 2, 3, 4]
ground = 'processed_mirrored_infidelities.npz'
preds = 'processed_cnn_mirrored_predictions.npz'
metric = 'r'

predictions = [np.load(pred_path.format(exp) + preds)['predictions'] for exp in exps]
infidelities = [np.load(sps_path.format(exp) + ground)['infidelities'] for exp in exps]

metrics = [metric_functions[metric](p, i) for p,i in zip(predictions, infidelities)]
metrics, np.mean(metrics), np.std(metrics)

# Circuit statistics

Random circuit depths

exps = [0, 1, 2, 3, 4]
exp_path = './experiment_{}/'
df_path = exp_path + '/simulation_results/high-fidelity-dataframe.csv'

def aggregate_data(metric: str) -> dict:
    results = {}
    for exp in exps:
        df = pd.read_csv(df_path.format(exp))
        min_metric, max_metric = min(df[metric]), max(df[metric])
        results[exp] = {'min': min_metric, 'max': max_metric}
    return results

depth_results = aggregate_data('F:Depth')
depth_results

Mirror circuit depths

exps = [0, 1, 2, 3, 4]
exp_path = './experiment_{}/'
df_path = exp_path + '/simulation_results/mirrored_dataframe.csv'

m_depth_results = aggregate_data('F:Depth')
m_depth_results

# Model size

from keras.models import load_model
from keras.optimizers import Adam
from keras.losses import MeanSquaredError

import json

from pygsti.extras import ml

exp_num = 1

model_path = f'./experiment_{exp_num}/models/'
exp_path = f'./experiment_{exp_num}/'

with open(exp_path + '/meta.json', 'r') as f:
    meta = json.load(f)

num_qubits = meta['num_qubits']
max_error_weight = meta['max_error_weight']
num_channels = meta['num_channels']
num_hops = meta['num_hops']
geometry = meta['geometry']

if meta['geometry'] == 'ring':
    adj_matrix = ml.newtools.ring_adj_matrix(num_qubits)
elif meta['geometry'] == 'algiers-t-bar':
    adj_matrix = ml.newtools.algiers_t_bar_adj_matrix()
elif meta['geometry'] == 't-bar':
    adj_matrix = ml.newtools.t_bar_adj_matrix()
elif meta['geometry'] == 'bowtie':
    adj_matrix = ml.newtools.bowtie_adj_matrix()
elif meta['geometry'] == 'melbourne':
    adj_matrix = ml.newtools.melbourne_adj_matrix()
else:
    print(f'You have not implemented the {geometry} geometry.')

laplace = ml.newtools.laplace_from_qubit_graph(adj_matrix)

error_gens = ml.newtools.up_to_weight_k_error_gens_from_qubit_graph(max_error_weight, num_qubits, laplace, num_hops)
error_gen_indices = [ml.newtools.error_gen_to_index(error[0], error[1]) for error in error_gens]


layer_snipper = ml.newneuralnets.layer_snipper_from_qubit_graph
layer_snipper_args = [num_qubits, num_channels, laplace, num_hops]

optimizer = Adam(learning_rate=1e-3)

circuits = np.load(exp_path+'/processed_inputs_and_outputs/processed_high_fidelity_circuits.npz')
infidelities = np.load(exp_path+'/processed_inputs_and_outputs/processed_infidelities.npz')

circuit_dense1 = ml.newneuralnets.CircuitErrorVec(num_qubits, num_channels, error_gens, layer_snipper,
                                                layer_snipper_args)

circuit_dense1.predict(circuits['train'][0:1])

circuit_dense1.summary()

# circuit_dense.load_weights(model_path.format(device) + 'custom_neural_network')
